requirements/requirements.txt
amqp==5.0.1
asgiref==3.2.10
billiard==3.6.3.0
boto3==1.15.15
botocore==1.18.15
celery==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.2
django-environ==0.4.5
django-extensions==3.0.9
django-filter==2.4.0
django-storages==1.10.1
djangorestframework==3.12.1
djangorestframework-bulk==0.2.1
idna==2.10
install==1.3.4
jmespath==0.10.0
kombu==5.0.2
Markdown==3.3
Pillow==7.2.0
prompt-toolkit==3.0.7
psycopg2-binary==2.8.6
pycurl==7.43.0.5
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
